Add files via upload
The project consist on a the implementation of four basic beats, each assigned to different switches on the Nexys A7 board. The four patterns are: single black beat, crotchets, quavers and triplets. With the previously mentioned beats, it is possible to make combinations in order to create more complex rhythms. 
The button C, works as a reset for all rhythms selected to start at the same time.